Fix: GNU mktemp requires at least three literal "X" in template
The script assumes BSD mktemp is installed on the local machine (eg. on *BSD
and MacOSX systems). There is a difference between BSD-derived mktemp and GNU
mktemp. If a template is specified, GNU mktemp requires
a block of at least three consecutive literal `X` characters to be present
in the template.  BSD mktemp doesn't have this requirement. Additionally,
`-t` option is marked as deprecated in GNU mktemp.

* Try to make the script work on GNU/Linux and BSD based systems and make
it a bit more cross-platform.
* Check that the temporary vagrant directory is created successfully and
  complain if its creation failed.